Refactored reuleauxArcTo function to fix a bug.
The bug occurs when the arc must go around and through each of the three
focal points. It was short circuiting because the logic looked for when
the current focal point to go through was the last focal point... which
are equal when the end point is behind the start point on the same arc.

Also, I fixed a bug where the circleContainsReuleaux was returning the
wrong result. It had originally just looked for a circle that contains
the corner points, however that's not enough. I just added an extra
check to look for the containment of all three corners AND that the
circle does not intersect the Reuleaux. This is enough to prove that the
circle contains the Reuleaux.